+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.0.1
+
+* Add test to ensure that every symbol from the Magic: The Gathering comprehensive rules has a corresponding item in `mtgSymbology`
+    * The source of truth for these symbols is https://api.scryfall.com/symbology
+
 ### 1.0.0
 
 * Initial stable release